+++ src/PluginMetadata.cs
@@ -4,5 +4,5 @@ internal class PluginMetadata
 {
     public const string PLUGIN_GUID = "umno.LethalCredit";
     public const string PLUGIN_NAME = "LethalCredit";
-    public const string PLUGIN_VERSION = "0.1.0";
+    public const string PLUGIN_VERSION = "0.1.1";
 }